13. Przeniesienie klasy Node do wnetrza IntLinkedList
