+++ src/main/java/edu/kis/vh/nursery/list/IntLinkedList.java
@@ -1,6 +1,35 @@
 package edu.kis.vh.nursery.list;
 
 public class IntLinkedList {
+    private static class Node {
+        private int value;
+        private Node prev;
+        private Node next;
+
+        public Node(int i) {
+            this.value = i;
+        }
+
+        public int getValue() {
+            return value;
+        }
+
+        public Node getPrev() {
+            return prev;
+        }
+
+        public void setPrev(Node prev) {
+            this.prev = prev;
+        }
+
+        public Node getNext() {
+            return next;
+        }
+
+        public void setNext(Node next) {
+            this.next = next;
+        }
+    }
 
     private static final int EMPTY_SIZE = -1;
     private Node last;
